requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
gunicorn>=20.1.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
